Adjusted dpvar getequation for Mineq case.
For matrix-valued dpvar expressions, "getequation" determines a set of monomials for each term of the matrix, discarding monomials that do not contribute. However, when a constraint of type 'ineq' is imposed, the resulting "prog.expr" format produced by "getequation" does not allow for (easy) processing/conversion of the constraint.  The function has been adjusted to take an optional input 'ineq', so that when this option is passed, a single set of unique monomials is returned to represent all elements of the matrix-valued expression.